feat: Menu 와 Restaurant 도메인 오류 처리

+++ src/main/java/christmas/domain/menuImpl/Appetizer.java
@@ -3,7 +3,7 @@ package christmas.domain.menuImpl;
 import christmas.domain.Menu;
 
 public class Appetizer extends Menu {
-    public Appetizer(String menuName, Integer price) {
+    public Appetizer(String menuName, long price) {
         super(menuName, price);
     }
 }
