+++ src/main/java/christmas/domain/menuImpl/Beverage.java
@@ -3,7 +3,7 @@ package christmas.domain.menuImpl;
 import christmas.domain.Menu;
 
 public class Beverage extends Menu {
-    public Beverage(String menuName, Integer price) {
+    public Beverage(String menuName, long price) {
         super(menuName, price);
     }
 }
