+++ src/main/java/christmas/domain/menuImpl/Dessert.java
@@ -3,7 +3,7 @@ package christmas.domain.menuImpl;
 import christmas.domain.Menu;
 
 public class Dessert extends Menu {
-    public Dessert(String menuName, Integer price) {
+    public Dessert(String menuName, long price) {
         super(menuName, price);
     }
 }
