+++ src/main/java/christmas/domain/menuImpl/Main.java
@@ -3,7 +3,7 @@ package christmas.domain.menuImpl;
 import christmas.domain.Menu;
 
 public class Main extends Menu {
-    public Main(String menuName, Integer price) {
+    public Main(String menuName, long price) {
         super(menuName, price);
     }
 }
